requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.128.8
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.20
requests==2.31.0
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.1.0
uvicorn==0.39.0

sqlalchemy
psycopg2-binary
pysui==0.59.0
cryptography

# Testing
pytest
httpx
